LayoutTests/fast/viewport/viewport-78-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "a" not recognized and ignored.
CONSOLE MESSAGE: line 3: Viewport argument key "1" not recognized and ignored.
viewport size 100x110 scale 3.200000 with limits [3.200000, 5.000000]
